# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: ChaloTogether API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings/{id}/accept:
    post:
      operationId: acceptBooking
      tags: [bookings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/reject:
    post:
      operationId: rejectBooking
      tags: [bookings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bookings/{id}/verify-otp:
    post:
      operationId: verifyRideOtp
      tags: [bookings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/complete:
    post:
      operationId: completeBooking
      tags: [bookings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Chat ────────────────────────────────────────────────────────────────────
...
  /chat/messages:
    get:
      operationId: getMessages
      tags: [chat]
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /chat/send:
    post:
      operationId: sendMessage
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /chat/read:
    post:
      operationId: markMessagesRead
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadInput"
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Trusted Contacts ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        demoToken:
          type: ["string", "null"]
        demoOtp:
          type: ["string", "null"]
...
    Booking:
      type: object
      required: [id, rideId, passengerId, status, pickupPoint, otp, createdAt]
      properties:
        id:
          type: integer
        rideId:
          type: integer
        passengerId:
          type: integer
        passengerName:
          type: string
        status:
          type: string
          enum: [requested, accepted, rejected, picked_up, completed, cancelled]
        pickupPoint:
          type: string
        farePerSeat:
          type: number
        otp:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    OtpInput:
      type: object
      required: [otp]
      properties:
        otp:
          type: string
...
    Message:
      type: object
      required: [id, senderId, receiverId, content, isRead, createdAt]
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        receiverId:
          type: integer
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [receiverId, content]
      properties:
        receiverId:
          type: integer
        content:
          type: string
    MarkReadInput:
      type: object
      required: [senderId]
      properties:
        senderId:
          type: integer
...