# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ChaloTogether API specification
 * OpenAPI spec version: 0.2.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  rideId: number;
  passengerId: number;
  passengerName?: string;
  status: BookingStatus;
  pickupPoint: string;
  farePerSeat?: number;
  otp: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}