# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ChaloTogether API specification
 * OpenAPI spec version: 0.2.0
 */

export interface SuccessResponse {
  message: string;
  /** @nullable */
  demoToken?: string | null;
  /** @nullable */
  demoOtp?: string | null;
}